+++ vale-sem-fome/src/components/Generics/Tables/TableDesktop/table-desktop.scss
@@ -1,0 +1,38 @@
+@import './../../../../assets/css/components/variables.scss';
+
+@media only screen and (min-width: 768px){
+    .table{
+        margin-top: 15px;
+        border-collapse: collapse;
+        table-layout: auto;
+        width: 100%;
+        
+        
+        thead{
+            text-align: left;
+            border-bottom: 1px solid $table-border-color;
+        }
+    
+        tr{
+            border-bottom: 1px solid $table-border-color;
+        }
+    
+        td, th{
+            height: 50px;
+            vertical-align: middle;
+            font-size: 1em;
+        }
+    
+        td:nth-child(1){
+            
+            font-weight: $font-weight-strong;
+        }
+    
+        td:nth-child(2), 
+        td:nth-child(3), 
+        td:nth-child(4){
+            color: $secundary-font-text-color;
+        }
+        
+    }
+}
