+++ vale-sem-fome/src/pages/Deprecated/Old/Home/messagens.txt
