--- vale-sem-fome/src/pages/Recipients/ShowOne/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function ShowOne() {
-    return (
-        <div>
-            <h1>Show One</h1>
-        </div>
-    )
-}
-
-export default ShowOne;
