+++ vale-sem-fome/src/pages/Recipients/View/FindByOne/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function RecipientsFindByOne() {
+    return (
+        <div>
+            <h1>Show One</h1>
+        </div>
+    )
+}
+
+export default RecipientsFindByOne;
