+++ vale-sem-fome/src/pages/index.js
@@ -1,19 +1,12 @@
-export { default as Home } from './Home'
-export { default as Obrigado } from './Obrigado'
-export { default as Voluntarios } from './Voluntarios'
-export { default as ObrigadoVoluntario } from './Voluntario/Obrigado'
+export { default as RecipientsForm } from './Recipients/Form'
+export { default as RecipientsThanks } from './Recipients/Others/Thanks'
+export { default as VolunteersForm } from './Volunteers/Form'
+export { default as VolunteersThanks } from './Volunteers/Others/Thanks'
 export { default as PageNotFound } from './PageNotFound'
-
-
-/**
- * Página de Teste
- */
-
-export { default as TestNewComponents } from './TestNewComponents'
 
 /**
  *  Views
  */
 
-export { default as AllRecipients } from './Recipients/ShowAll'
-export { default as OneRecipient } from './Recipients/ShowOne'
+export { default as RecipientsFindAll } from './Recipients/View/FindAll'
+export { default as RecipientsFindByOne } from './Recipients/View/FindByOne'
